service layer refactor ended (but I didn't test it yet)
