+++ src/main/java/com/restapi/vinted/service/impl/ClothesServiceimpl.java
@@ -6,10 +6,11 @@ import com.restapi.vinted.exception.ApiException;
 import com.restapi.vinted.exception.ResourceNotFoundException;
 import com.restapi.vinted.payload.ClotheDto;
 import com.restapi.vinted.payload.ClotheResponse;
+import com.restapi.vinted.repository.CategoryRepository;
 import com.restapi.vinted.repository.ClotheRepository;
 import com.restapi.vinted.repository.UserRepository;
 import com.restapi.vinted.service.ImageService;
-import com.restapi.vinted.service.MyClothesService;
+import com.restapi.vinted.service.ClothesService;
 import com.restapi.vinted.utils.Constant;
 import org.modelmapper.ModelMapper;
 
@@ -27,22 +28,54 @@ import java.util.List;
 
 @Service
 @PreAuthorize("hasAnyRole('ADMIN', 'USER')")
-public class MyClothesServiceimpl implements MyClothesService {
+public class ClothesServiceimpl implements ClothesService {
 
     private final ClotheRepository clotheRepository;
     private final ImageService imageService;
     private final ModelMapper mapper;
     private final UserRepository userRepository;
+    private final CategoryRepository categoryRepository;
 
-    public MyClothesServiceimpl(ClotheRepository clotheRepository, ImageService imageService,
-                                ModelMapper mapper, UserRepository userRepository) {
+
+    public ClothesServiceimpl(ClotheRepository clotheRepository, ImageService imageService,
+                              ModelMapper mapper, UserRepository userRepository, CategoryRepository categoryRepository) {
         this.clotheRepository = clotheRepository;
         this.imageService = imageService;
         this.mapper = mapper;
         this.userRepository = userRepository;
+        this.categoryRepository = categoryRepository;
     }
 
 
+    @Override
+    public ClotheResponse getAllClothesByCategory(long categoryId, int pageNo, int pageSize,
+                                                  String sortBy, String direction){
+
+        //create Sort, and Page object
+        Sort sort = direction.equalsIgnoreCase(Sort.Direction.ASC.name()) ?
+                Sort.by(sortBy).ascending() : Sort.by(sortBy).descending();
+
+        Pageable page = PageRequest.of(pageNo, pageSize, sort);
+
+        //create Page<Clothe> with custom DB method
+        categoryRepository.findById(categoryId)
+                .orElseThrow(() -> new ResourceNotFoundException("Category", "id", categoryId));
+
+        Page<Clothe> clothes = clotheRepository.findByCategoryId(categoryId, page);
+
+        return getClotheResponse(pageNo, pageSize, clothes);
+    }
+
+    @Override
+    public ClotheDto getClotheById(long clotheId) {
+
+        Clothe clothe = clotheRepository.findById(clotheId)
+                .orElseThrow( ()-> new ResourceNotFoundException("Clothe", "id", clotheId));
+
+        return mapToDto(clothe);
+    }
+    
+    // only logged-in user action
     @Override
     @Transactional
     public ClotheDto addClothe(ClotheDto clotheDto, List<MultipartFile> images, String email) {
@@ -58,6 +91,7 @@ public class MyClothesServiceimpl implements MyClothesService {
         return mapToDto(savedClothe);
     }
 
+    // only logged-in user action
     @Override
     @Transactional(readOnly = true)
     public ClotheResponse getMyClothes(int pageNo, int pageSize, String sortBy,
@@ -74,20 +108,12 @@ public class MyClothesServiceimpl implements MyClothesService {
         //getting page of clothes owned by logged-in user
         Page<Clothe> clothes = clotheRepository.findByUserId(user.getId(), page);
 
-        //creating ClotheResponse - bring more information about clothes to a client
-        ClotheResponse clotheResponse = new ClotheResponse();
-
-        clotheResponse.setClothes(clothes.stream().map(this::mapToDto).toList());
-        clotheResponse.setPageNo(pageNo);
-        clotheResponse.setTotalPages(clothes.getTotalPages());
-        clotheResponse.setPageSize(pageSize);
-        clotheResponse.setLast(clothes.isLast());
-
-        return clotheResponse;
+        return getClotheResponse(pageNo, pageSize, clothes);
     }
 
+    // owner action
     @Override
-    @PreAuthorize("@myClothesServiceimpl.isOwner(#id, #email)")
+    @PreAuthorize("@clothesServiceimpl.isOwner(#id, #email)")
     @Transactional
     public ClotheDto updateClothe(long id, ClotheDto clotheDto,
                                   List<MultipartFile> newImages, List<String> deletedImages, String email) {
@@ -99,26 +125,15 @@ public class MyClothesServiceimpl implements MyClothesService {
         clothe.setPrice(clotheDto.getPrice());
         clothe.setSize(clotheDto.getSize());
 
-        if (newImages != null && !newImages.isEmpty()) {
-            var newImageNames = newImages.stream().map(imageService::saveImage).toList();
-            clothe.getImages().addAll(newImageNames);
-        }
-
-        if (deletedImages != null && !deletedImages.isEmpty()) {
-            clothe.getImages().removeAll(deletedImages);
-            deletedImages.forEach(imageService::deleteImage);
-        }
-
-        if(clothe.getImages().size() > 5)
-            throw new ApiException(HttpStatus.BAD_REQUEST, Constant.IMAGES_VALIDATION_FAILED);
+        updateImages(clothe, newImages, deletedImages);
 
         Clothe updatedClothe = clotheRepository.save(clothe);
         return mapToDto(updatedClothe);
     }
 
-
+    // owner action
     @Override
-    @PreAuthorize("@myClothesServiceimpl.isOwner(#id, #email)")
+    @PreAuthorize("@clothesServiceimpl.isOwner(#id, #email)")
     @Transactional
     public String deleteClothe(long id, String email) {
 
@@ -129,14 +144,6 @@ public class MyClothesServiceimpl implements MyClothesService {
         return "Clothe deleted successfully!";
     }
 
-
-
-    private Clothe getClotheFromDB(long id){
-        return clotheRepository.findById(id).orElseThrow(
-                () -> new ResourceNotFoundException("Clothe", "id", id)
-        );
-    }
-
     @Transactional(readOnly = true)
     public boolean isOwner(long clotheId, String email){
         var clothe = getClotheFromDB(clotheId);
@@ -144,6 +151,27 @@ public class MyClothesServiceimpl implements MyClothesService {
         if (!clothe.getUser().equals(user))
             throw new ApiException(HttpStatus.FORBIDDEN, Constant.NOT_OWNER);
         return true;
+    }
+
+    private void updateImages(Clothe clothe, List<MultipartFile> newImages, List<String> deletedImages) {
+        if (deletedImages != null && !deletedImages.isEmpty()) {
+            clothe.getImages().removeAll(deletedImages);
+            deletedImages.forEach(imageService::deleteImage);
+        }
+
+        if(clothe.getImages().size() + newImages.size() > 5)
+            throw new ApiException(HttpStatus.BAD_REQUEST, Constant.IMAGES_VALIDATION_FAILED);
+
+        if (newImages != null && !newImages.isEmpty()) {
+            var newImageNames = newImages.stream().map(imageService::saveImage).toList();
+            clothe.getImages().addAll(newImageNames);
+        }
+    }
+
+    private Clothe getClotheFromDB(long id){
+        return clotheRepository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Clothe", "id", id)
+        );
     }
 
     private User getUser(String email){
@@ -157,5 +185,16 @@ public class MyClothesServiceimpl implements MyClothesService {
 
     private ClotheDto mapToDto(Clothe clothe){
         return mapper.map(clothe, ClotheDto.class);
+    }
+
+    private ClotheResponse getClotheResponse(int pageNo, int pageSize, Page<Clothe> clothes){
+        ClotheResponse clotheResponse = new ClotheResponse();
+
+        clotheResponse.setClothes(clothes.stream().map(this::mapToDto).toList());
+        clotheResponse.setPageNo(pageNo);
+        clotheResponse.setTotalPages(clothes.getTotalPages());
+        clotheResponse.setPageSize(pageSize);
+        clotheResponse.setLast(clothes.isLast());
+        return clotheResponse;
     }
 }
